// src/app/core/clinica-tipo-especialidade-dto.ts
export class ClinicaTipoEspecialidadeDto {                           
	idClinica: number;
	nomeRazaoSocial: string;
	email: string;
	cnpj: string;
	nomeFantasia: string;
	ativo:boolean;
	dataCadastro: Date;
	uf: string;
	cidade: string;
	bairro: string;   
    endereco: string;
	telefone01: string;
	telefone02: string;
	celular: string;
	descricaoEspecialidade: string;
	                        
}        
                      
        